2024/04/09 v2.2 RedPacket： 1.构建日志文件输出。 2.日志实现使用log4j2替换logback，启用全量异步模式。     需要启动参数-DLog4jContextSelector=org.apache.logging.log4j.core.async.AsyncLoggerContextSelector 	或配置文件log4j2.component.properties中Log4jContextSelector=org.apache.logging.log4j.core.async.AsyncLoggerContextSelector 3.重构业务日志格式，用红包key和用户ID作为前缀，提高可读性。 4.统一异常处理响应时，捕获的未处理异常按照栈格式输出。 5.新增通过配置控制日志输出的功能，包括控制台、文件、大数据接口（Kafka） 	a) 新增自定义日志级别：业务和大数据。 	b) 日志输出到文件支持错误日志、业务日志、普通日志分离，分别使用不同的生成和保留规则。 	c) 大数据日志提供按字节编解码方法，并启用惰性日志。 6.发起抢红包后、写入缓存前、参与抢红包后的扩展方法在组合类中逐个进行异常捕获。

+++ RedPacket/src/main/java/com/practice/common/result/PublishResult.java
@@ -10,16 +10,16 @@ public class PublishResult {
     private final String dateTime; // 发起抢红包的日期时间
     private final int amount; // 发起抢红包的总金额
     private final int shareNum; // 发起抢红包的份数
-    private final long expireTime; // 发起抢红包的有效期
+    private final int expireTime; // 发起抢红包的有效期
 
     /**
      * 发起抢红包成功，返回发起日期时间、金额、份数、有效期
      */
-    public static PublishResult publishSuccess(String dateTime, int amount, int shareNum, long expireTime) {
+    public static PublishResult publishSuccess(String dateTime, int amount, int shareNum, int expireTime) {
         return new PublishResult(dateTime, amount, shareNum, expireTime);
     }
 
-    public PublishResult(String dateTime, int amount, int shareNum, long expireTime) {
+    public PublishResult(String dateTime, int amount, int shareNum, int expireTime) {
         this.dateTime = dateTime;
         this.amount = amount;
         this.shareNum = shareNum;
